requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
youtube-transcript-api==0.6.2
langchain==0.1.4
langchain-openai==0.0.5
langchain-huggingface==0.0.1
langchain-community==0.0.17
faiss-cpu==1.7.4
sentence-transformers==2.3.1